db-engine/requirements.txt
httpx==0.24.1
psycopg2-binary==2.9.9
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
pydantic==2.5.2